// src/app/store/models/driver.model.ts
export interface Driver {
  TempID: string;
  BrokerType: string;
  FirstName: string;
  MiddleName: string;
  LastName: string;
  CompanyName: string;
  StreetName: string;
  Unit: string;
  City: string;
  Country: string;
  Province: string;
  PostalCode: string;
  Email: string;
  AlternateEmail: string;
  HomePhone: string;
  CellPhone: string;
  EmergencyContactName: string;
  EmergencyPhone: string;
  Fax: string;
  AlternateFax: string;
  SIN: string;
  BIN: string;
  ScanID: string;
  Charges: boolean | null;
  Conviction: boolean | null;
  Gender: string;
  BrokerRegion: number;
  StartDate: Date;
  VehicleUseRequired: boolean | null;
  BCRequired: boolean | null;
  DARequired: boolean | null;
  LeaseDocumentsRequired: boolean | null;
  ProceedWithApplicant: boolean | null;
  GST: string;
  DOB: Date;
  SpecificReason: string;
  ReasonForRejection: string;
  ApprovingRecruiterId: string;
  ApprovingRecruiterEmail: string;
  VehicleType: string;
  LicenseNumber: string;
  LicenseIssueDate: Date;
  LicenseExpiryDate: Date;
  VehicleMake: string;
  VehicleModel: string;
  VehicleYear: string;
  VehicleInsurance: string;
  InsurancePolicyNo: string;
  InsurancePolicyExpiryDate: Date;
  LicensePlateNo: string;
  LicensePlateExpiryDate: Date;
  VehicleVinNo: string;
  VehicleLeased: boolean | null;
}


export interface DriverTemp {
  TempID: string;
  Charges: boolean | null;
  Conviction:boolean | null;
  VehicleUseRequired: boolean | null;
  BCRequired: boolean | null;
  DARequired : boolean | null;
  LeaseDocumentsRequired: boolean | null;
  ProceedWithApplicant: boolean | null;
  BrokerRegion:string;
  BrokerType: string;
  StartDate:Date;  
  FirstName: string;
  MiddleName: string;
  LastName: string;  
  DOB:Date;
  Email: string;  
  CellPhone: string;
  ValidWorkPermit:boolean | null;
  ReasonForRejection:string;  
  ApprovingRecruiterId: string;
  ApprovingRecruiterEmail: string;    
}